iceberg_lake/requirements_kafka.txt
kafka-python>=2.0.2
jaydebeapi>=1.2.3
jpype1>=1.3.0